+++ src/424select.js
@@ -390,10 +390,11 @@ yy.Select.prototype.compileSelectGroup2 = function(query) {
 	var s = query.selectgfns;
 //	console.log('selectg:',s);
 	return new Function('g,params,alasql','var y;'+s+'return r');
-}
+};
 
 // SELECY * REMOVE [COLUMNS] col-list, LIKE ''
 yy.Select.prototype.compileRemoveColumns = function(query) {
+	var self = this;
 	if(typeof this.removecolumns !== 'undefined') {
 		query.removeKeys = query.removeKeys.concat(
 			this.removecolumns.filter(function (column) {
@@ -404,7 +405,9 @@ yy.Select.prototype.compileRemoveColumns = function(query) {
 		query.removeLikeKeys = this.removecolumns.filter(function (column) {
 				return (typeof column.like !== 'undefined');
 			}).map(function(column){
-				return new RegExp((column.like.value||'').replace(/\%/g,'.*').replace(/\?|_/g,'.'),'g');
+//				return new RegExp((column.like.value||'').replace(/\%/g,'.*').replace(/\?|_/g,'.'),'g');
+				return column.like.value;
 			});
 	}
-}
+};
+