+++ test/test366.js
@@ -34,12 +34,16 @@ describe('Test 366 wildcards for LIKE', function() {
 						done();
 			  });
 
-			  it.skip('can escape wildcards',function(done){
-			    	assert.equal(1,alasql(sql+"'\\%'",[data]));
+			  it('can escape wildcards',function(done){
+			  		// AG: Changed to proper escape character
+			    	assert.equal(1,alasql(sql+"'_!%_' ESCAPE '!'",[data]));
+//			    	assert.equal(1,alasql(sql+"'\\%'",[data]));
 			  		done();
 				});
 
 			  it('Finds prepending',function(done){
+
+//			  	console.log(alasql(sql+"'%a'",[data]));
 			    	assert.equal(0,alasql(sql+"'%a'",[data]));
 			    	assert.equal(6,alasql(sql+"'%b'",[data]));
 			  		done();
@@ -59,49 +63,51 @@ describe('Test 366 wildcards for LIKE', function() {
 						done();
 				});
 
-
 		});
 
 		describe('with ?', function() {
 
 				it('find n long elements',function(done){
-						assert.equal(1, alasql(sql+"'??'",[data]));
+					// I changed from ? to _
+						assert.equal(1, alasql(sql+"'__'",[data]));
 						done();
 				});
 
 				it('finds none',function(done){
-						assert.equal(0, alasql(sql+"'?q'",[data]));
+					// I changed from ? to _
+						assert.equal(0, alasql(sql+"'_q'",[data]));
 						done();
 				});
 
-				it.skip('can escape wildcards',function(done){
-						assert.equal(1,alasql(sql+"'\\?'",[data]));
+				it('can escape wildcards',function(done){
+					// Changed escape character from // to ! and ? to _
+						assert.equal(0,alasql(sql+"'!__' ESCAPE '!'",[data]));
+//						assert.equal(1,alasql(sql+"'!_' ESCAPE '!'",[data]));
 						done();
 				});
 
 				it('Finds prepending',function(done){
-						assert.equal(0,alasql(sql+"'?a'",[data]));
-						assert.equal(1,alasql(sql+"'?b'",[data]));
-						assert.equal(5,alasql(sql+"'??b'",[data]));
+						assert.equal(0,alasql(sql+"'_a'",[data]));
+						assert.equal(1,alasql(sql+"'_b'",[data]));
+						assert.equal(5,alasql(sql+"'__b'",[data]));
 						done();
 				});
 
 
 				it('Finds center',function(done){		// Not supported yet
-						assert.equal(1,alasql(sql+"'?a?'",[data]));
-						assert.equal(1,alasql(sql+"'?b?'",[data]));
+						assert.equal(1,alasql(sql+"'_a_'",[data]));
+						assert.equal(1,alasql(sql+"'_b_'",[data]));
 						done();
 				});
 
 
 				it('Finds postpending',function(done){		// Not supported yet
-						assert.equal(1,alasql(sql+"'a?'",[data]));
-						assert.equal(0,alasql(sql+"'b?'",[data]));
+						assert.equal(1,alasql(sql+"'a_'",[data]));
+						assert.equal(0,alasql(sql+"'b_'",[data]));
 						done();
 				});
 
 		});
-
 
 		describe('with _', function() {
 
@@ -115,8 +121,8 @@ describe('Test 366 wildcards for LIKE', function() {
 						done();
 				});
 
-				it.skip('can escape wildcards',function(done){
-						assert.equal(1,alasql(sql+"'\\_'",[data]));
+				it('can escape wildcards',function(done){
+						assert.equal(1,alasql(sql+"'_!__' ESCAPE '!'",[data]));
 						done();
 				});
 
@@ -145,6 +151,7 @@ describe('Test 366 wildcards for LIKE', function() {
 
 
 
+
 				/*
 				{x:'ab'}
 				,{x:'-ab'}
@@ -157,3 +164,4 @@ describe('Test 366 wildcards for LIKE', function() {
 
 
 });
+