+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * Check https://trello.com/b/qxz65pVi/alasql-roadmap for roadmap
 * Added SUBSTR() alias for MID() function (for SQLite compatibility)
-
+* Changed LIKE function to alasql.utils.like
+* Added LIKE ESCAPE functionality
 
 ## Version 0.2.0 "Athens" (13.07.2015)
 The purpose of this release were hard work on: